project/src/store/user-process/user-process.tsx
import {createSlice} from '@reduxjs/toolkit';
import {NameSpace, AuthorizationStatus} from '../../const';
import {UserProcess} from '../../types/state';

const initialState: UserProcess = {
  authorizationStatus: AuthorizationStatus.Unknown,
  avatarUrl: '',
};

export const userProcess = createSlice({
  name: NameSpace.User,
  initialState,
  reducers: {
    requireAuthorization: (state, action) => {
      state.authorizationStatus = action.payload;
    },
    loadUserAvatarUrl: (state, action) => {
      state.avatarUrl = action.payload;
    },
  },
});

export const {requireAuthorization, loadUserAvatarUrl} = userProcess.actions;
